requirements.txt
Django==5.1.2
djangorestframework==3.15.2
dj-database-url==2.2.0
psycopg[binary]==3.2.10
gunicorn==23.0.0
whitenoise==6.7.0
django-cors-headers==4.4.0
